requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.24.0
python-multipart==0.0.6
deepface==0.0.79
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
tensorflow-intel==2.13.0
typing-extensions==4.5.0
beautifulsoup4==4.12.3
